[ROCm][Perf] MiniMax-M3 MXFP4: indexer shared-K-reuse + cnb hoist, bf16 router gate, sparse num_stages=1
Stacks on #46474 (MoE fusion). Aligns vLLM to ATOM on MI355X/gfx950, TP4, bf16 KV:
- index_topk: lightning-indexer re-read the shared prefix K once PER REQUEST; reuse
  the cross-request shared prefix K once. decode _decode_index_score 4.65 -> 1.37 ms/step.
- indexer: hoist common_num_blocks to once-per-step (was 57x/step). decode per-step
  23.97 -> 21.1 ms, TPOT 23.86 -> 21.16 (beats ATOM).
- amd/model.py: router gate GEMM in bf16 (out_dtype stays fp32). prefill 684 -> 658 ms.
- sparse_attn: num_stages=1 on gfx950 too (match ATOM). prefill 804 -> 705 ms.

Net: decode TPOT 21.45 ms (-0.2% vs atom); prefill +1.6% (with INT4 AR + aiter
AITER_FP4_FORCE_1STAGE env). Outputs verified coherent.

+++ vllm/models/minimax_m3/amd/model.py
@@ -314,13 +314,14 @@ class MiniMaxM3MoE(nn.Module):
         else:
             self.e_score_correction_bias = None
 
-        # Router weights are stored in fp32; GateLinear upcasts the bf16
-        # activations and computes the gate in fp32 (fp32 router logits).
+        # Router gate GEMM runs in bf16 (matches ATOM); out_dtype stays fp32 so
+        # routing logits keep fp32 precision. Avoids the fp32 activation upcast on
+        # ROCm/gfx950 (prefill per-step ~684 -> ~658 ms, accuracy unchanged).
         self.gate = GateLinear(
             config.hidden_size,
             config.num_local_experts,
             bias=False,
-            params_dtype=torch.float32,
+            params_dtype=torch.bfloat16,
             out_dtype=torch.float32,
             prefix=f"{prefix}.gate",
         )
